+++ src/main/java/io/github/astrapi69/swing/tree/panel/TreeNodeJTreeElementPanel.java
@@ -27,7 +27,7 @@ package io.github.astrapi69.swing.tree.panel;
 import io.github.astrapi69.gen.tree.TreeNode;
 import io.github.astrapi69.model.BaseModel;
 import io.github.astrapi69.model.api.IModel;
-import io.github.astrapi69.swing.tree.JTreeElement;
+import io.github.astrapi69.swing.renderer.tree.JTreeElement;
 
 /**
  * The abstract class {@link TreeNodeJTreeElementPanel} a given {@link TreeNode} parameterized with