+++ src/main/java/module-info.java
@@ -1,0 +1,17 @@
+module swing.tree.component.main
+{
+	requires java.desktop;
+	requires data.api;
+	requires gen.tree;
+	requires lombok;
+	requires swing.renderer.main;
+	requires model.data;
+	requires swing.base.components.main;
+	requires swing.table.components.main;
+	requires menu.actions.main;
+	requires roboter.main;
+	requires component.model.main;
+	requires swing.model.components.main;
+	requires com.miglayout.core;
+	requires com.miglayout.swing;
+}