+++ src/test/java/io/github/astrapi69/swing/tree/panel/content/TestTreeNodeGenericTreeElementWithGenericContentPanel.java
@@ -31,7 +31,7 @@ import io.github.astrapi69.awt.window.adapter.CloseWindow;
 import io.github.astrapi69.gen.tree.TreeNode;
 import io.github.astrapi69.model.BaseModel;
 import io.github.astrapi69.model.api.IModel;
-import io.github.astrapi69.swing.tree.GenericTreeElement;
+import io.github.astrapi69.swing.renderer.tree.GenericTreeElement;
 import io.github.astrapi69.swing.tree.panel.JTreePanel;
 import io.github.astrapi69.test.instance.TestTreeNodeFactory;
 import io.github.astrapi69.test.object.Permission;